unified_face_analyzer/requirements.txt
# Core ML/CV Libraries
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Face Analysis
mediapipe>=0.10.0        # MediaPipe 468-point landmark detection
dlib>=19.24.0            # dlib 68-point landmark detection

# CLIP (for gender, glasses, beard classification)
git+https://github.com/openai/CLIP.git

# Transformers (for CLIP model loading)
transformers>=4.30.0

# GUI (optional)
tkinterdnd2-universal>=0.3.0

# Configuration & Utilities
pyyaml>=6.0
